backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
mlx>=0.9; platform_system == "Darwin"
